Shopping_Cart
The Advanced Shopping Cart Management System is a comprehensive e-commerce platform designed for online product sales. It uses a 3-tier architecture, with a backend relational database, object-oriented PHP, and a web browser for front-end management. The system aims to simplify and enhance the shopping experience for customers, streamline the purchasing process, ensure product delivery, and position Bangladesh as a global competitor. Key technologies include PHP for business logic, MySQL for database management, HTML for structural design, CSS for formatting, and JavaScript for form validation and animation. The system's graphical user interface (GUI) design focuses on user attraction and ease of use. Future plans include credit card validation, online transactions, and email notifications. The system also considers third-party software for Visa checks to enhance security. The project's commitment to simplicity, user-friendly design, and future enhancements positions it as a robust and adaptive solution in the dynamic landscape of online retail.

The project objectives are:
▪ To make shopping easier and more comfortable.
▪ To serve the customers without wasting their precious time.
▪ To reach the products at the customer&#39;s address with great care.
▪ To represent Bangladesh as a competitor in e-commerce.
▪ User-Friendly Interface.
▪ Integration with Inventory Management.
▪ Shopping Cart Functionality.